requirements.txt
django
djangorestframework
celery
django-celery-beat
redis
numpy
pandas
scipy
scikit-learn
torch
torchvision
spectral
opencv-python
shap
lime
captum
matplotlib
plotly
pillow
django-storages
